Add drag-and-drop functionality for canvas elements
Introduced a new drag-and-drop feature for different canvas elements. This includes new interfaces and properties for Page and Shape, renaming of methods and components (from 'page' to 'pages'), as well as the inclusion of the react-dnd library for the drag-and-drop functionality. Also added the ability to add shapes to the canvas.