Refactor drag and drop functionality in pages and elements
In the previous implementation, elements' positions were managed in the local state of PageComponent which caused redundancy. This commit refactors that by moving the elements' positions into the pages state in the usePages custom hook. As part of this change, the 'elementPositions' prop has been removed from the Elements component, and each Element's style is now directly derived from its 'position' attribute.